Extracted the ChangeUserPasswordSettingsCtrl as UserSettingsController in separate controller file. Extracted constants and utilities in separate files.
